Moved the removeCompleteFuture method to the IndexGenerator class.
